chore(astro): reorganize integrations hooks (#10896)
* chore(astro): reorganize integrations hooks

* fix: test paths
