+++ packages/astro/src/integrations/features-validation.ts
