+++ packages/astro/src/integrations/hooks.ts
@@ -20,7 +20,7 @@ import { buildClientDirectiveEntrypoint } from '../core/client-directive/index.j
 import { mergeConfig } from '../core/config/index.js';
 import type { AstroIntegrationLogger, Logger } from '../core/logger/core.js';
 import { isServerLikeOutput } from '../prerender/utils.js';
-import { validateSupportedFeatures } from './astroFeaturesValidation.js';
+import { validateSupportedFeatures } from './features-validation.js';
 
 async function withTakingALongTimeMsg<T>({
 	name,
