+++ packages/astro/test/units/integrations/api.test.js
@@ -1,7 +1,7 @@
 import * as assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
-import { validateSupportedFeatures } from '../../../dist/integrations/astroFeaturesValidation.js';
-import { runHookBuildSetup, runHookConfigSetup } from '../../../dist/integrations/index.js';
+import { validateSupportedFeatures } from '../../../dist/integrations/features-validation.js';
+import { runHookBuildSetup, runHookConfigSetup } from '../../../dist/integrations/hooks.js';
 import { defaultLogger } from '../test-utils.js';
 
 describe('Integration API', () => {
